requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.4.4
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.7
clip==1.0
contourpy==1.1.1
cycler==0.12.1
diffusers==0.27.2
einops==0.7.0
filelock==3.13.4
flask==3.0.3
flask-ngrok==0.0.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.2
huggingface-hub==0.22.2
idna==3.4
imageio==2.34.1
importlib-metadata==7.1.0
importlib-resources==6.4.0
invisible-watermark==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
latent-diffusion==0.0.1
lazy-loader==0.4
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.74
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.0.3
parso==0.8.4
pillow==10.2.0
pip==20.3.3
protobuf==5.26.1
pudb==2024.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.7.1
pydantic-core==2.18.2
pyDeprecate==0.3.2
pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
setuptools==68.2.2
six==1.16.0
sympy==1.12
taming-transformers==0.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
threadpoolctl==3.4.0
tifffile==2023.7.10
tokenizers==0.12.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==1.3.2
torchvision==0.12.0+cu113
tqdm==4.66.2
transformers==4.31.0
triton==2.3.0
typing-extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
urwid==2.6.11
urwid-readline==0.14
wcwidth==0.2.13
werkzeug==3.0.2
wheel==0.41.2
yarl==1.9.4
zipp==3.18.1
